requirements.txt
beautifulsoup4==4.13.4
mysql-connector-python==9.3.0
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
